rename: onboarding 컴포넌트 파일 위치 옮김
